TST: re-organize tests a bit
