+++ caproto/tests/conftest.py
@@ -15,9 +15,21 @@ import caproto.curio.client  # noqa
 import caproto.curio.server  # noqa
 import caproto.threading  # noqa
 import caproto.threading.client  # noqa
+from caproto.curio.server import find_next_tcp_port
+import caproto.curio.server as server
+
+
+str_alarm_status = ca.ChannelAlarm(
+    status=ca.AlarmStatus.READ,
+    severity=ca.AlarmSeverity.MINOR_ALARM,
+    alarm_string='alarm string',
+)
 
 
 _repeater_process = None
+
+REPEATER_PORT = 5065
+SERVER_HOST = '0.0.0.0'
 
 
 def start_repeater():
@@ -46,6 +58,66 @@ def stop_repeater():
     _repeater_process.wait()
     print('OK')
     _repeater_process = None
+
+
+@pytest.fixture(scope='function')
+def curio_server():
+    caget_pvdb = {
+        'pi': ca.ChannelDouble(value=3.14,
+                               lower_disp_limit=3.13,
+                               upper_disp_limit=3.15,
+                               lower_alarm_limit=3.12,
+                               upper_alarm_limit=3.16,
+                               lower_warning_limit=3.11,
+                               upper_warning_limit=3.17,
+                               lower_ctrl_limit=3.10,
+                               upper_ctrl_limit=3.18,
+                               precision=5,
+                               units='doodles',
+                               ),
+        'enum': ca.ChannelEnum(value='b',
+                               enum_strings=['a', 'b', 'c', 'd'],
+                               ),
+        'int': ca.ChannelInteger(value=33,
+                                 units='poodles',
+                                 lower_disp_limit=33,
+                                 upper_disp_limit=35,
+                                 lower_alarm_limit=32,
+                                 upper_alarm_limit=36,
+                                 lower_warning_limit=31,
+                                 upper_warning_limit=37,
+                                 lower_ctrl_limit=30,
+                                 upper_ctrl_limit=38,
+                                 ),
+        'char': ca.ChannelChar(value=b'3',
+                               units='poodles',
+                               lower_disp_limit=33,
+                               upper_disp_limit=35,
+                               lower_alarm_limit=32,
+                               upper_alarm_limit=36,
+                               lower_warning_limit=31,
+                               upper_warning_limit=37,
+                               lower_ctrl_limit=30,
+                               upper_ctrl_limit=38,
+                               ),
+        'str': ca.ChannelString(value='hello',
+                                alarm=str_alarm_status,
+                                reported_record_type='caproto'),
+        }
+
+    async def run_server():
+        port = find_next_tcp_port(host=SERVER_HOST)
+        print('Server will be on', (SERVER_HOST, port))
+        ctx = server.Context(SERVER_HOST, port, caget_pvdb, log_level='DEBUG')
+        try:
+            await ctx.run()
+        except Exception as ex:
+            print('Server failed', ex)
+            raise
+        finally:
+            print('Server exiting')
+
+    return run_server, caget_pvdb
 
 
 @pytest.fixture(scope='function')
